AlumniAPI/requirements.txt
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
channels==4.3.2
python-dotenv==1.2.1
psycopg2-binary==2.9.11
requests==2.32.5
gunicorn==23.0.0
